// protobuf/protoc-gen-tsvalidator/valtest/geo.validator.ts
// Code generated by protoc-gen-tsvalidator. DO NOT EDIT.
// versions:
// 	protoc-gen-tsvalidator v2.6.0
// 	protoc                 v4.24.4
// source: geo.proto

import { z } from "zod";

export const GeoValidationMessage_Latitude = z
  .number()
  .min(-90)
  .max(90)
  .optional();
export const GeoValidationMessage_Longitude = z
  .number()
  .min(-180)
  .max(180)
  .optional();
export const GeoValidationMessage_Accuracy = z
  .number()
  .int()
  .min(0)
  .max(10000)
  .optional();
export const GeoValidationMessageValidator = z.object({
  latitude: GeoValidationMessage_Latitude,
  longitude: GeoValidationMessage_Longitude,
  accuracy: GeoValidationMessage_Accuracy,
});